+++ core/jvm/src/main/scala/zio/RuntimeConfigPlatformSpecific.scala
@@ -21,22 +21,9 @@ import zio.internal.stacktracer.impl.AkkaLineNumbersTracer
 import zio.internal.tracing.TracingConfig
 import zio.internal.{Blocking, Tracing, ZLogger}
 
-import java.lang.ref.WeakReference
-import java.util.concurrent.ConcurrentHashMap
-import java.util.{Collections, Map => JMap, Set => JSet, WeakHashMap}
 import scala.concurrent.ExecutionContext
 
 private[zio] trait RuntimeConfigPlatformSpecific {
-
-  /**
-   * Adds a shutdown hook that executes the specified action on shutdown.
-   */
-  def addShutdownHook(action: () => Unit): Unit =
-    java.lang.Runtime.getRuntime.addShutdownHook {
-      new Thread {
-        override def run() = action()
-      }
-    }
 
   /**
    * A Runtime with settings suitable for benchmarks, specifically with Tracing
@@ -61,13 +48,6 @@ private[zio] trait RuntimeConfigPlatformSpecific {
    * yielding to other fibers.
    */
   final val defaultYieldOpCount = 2048
-
-  /**
-   * Returns the name of the thread group to which this thread belongs. This
-   * is a side-effecting method.
-   */
-  final def getCurrentThreadGroup: String =
-    Thread.currentThread.getThreadGroup.getName
 
   /**
    * A `RuntimeConfig` created from Scala's global execution context.
@@ -111,41 +91,9 @@ private[zio] trait RuntimeConfigPlatformSpecific {
     fromExecutor(Executor.fromExecutionContext(defaultYieldOpCount)(ec))
 
   /**
-   * Returns whether the current platform is ScalaJS.
-   */
-  val isJS = false
-
-  /**
-   * Returns whether the currently platform is the JVM.
-   */
-  val isJVM = true
-
-  /**
-   * Returns whether the currently platform is Scala Native.
-   */
-  val isNative = false
-
-  /**
    * Makes a new default runtime configuration. This is a side-effecting
    * method.
    */
   def makeDefault(yieldOpCount: Int = defaultYieldOpCount): RuntimeConfig =
     fromExecutor(Executor.makeDefault(yieldOpCount))
-
-  final def newWeakHashMap[A, B](): JMap[A, B] =
-    Collections.synchronizedMap(new WeakHashMap[A, B]())
-
-  final def newConcurrentWeakSet[A](): JSet[A] =
-    Collections.synchronizedSet(newWeakSet[A]())
-
-  final def newWeakSet[A](): JSet[A] =
-    Collections.newSetFromMap(new WeakHashMap[A, java.lang.Boolean]())
-
-  final def newConcurrentSet[A](): JSet[A] = ConcurrentHashMap.newKeySet[A]()
-
-  final def newWeakReference[A](value: A): () => A = {
-    val ref = new WeakReference[A](value)
-
-    () => ref.get()
-  }
 }
