+++ core/shared/src/main/scala/zio/internal/package.scala
@@ -22,8 +22,8 @@ package object internal {
 
   @deprecated("use RuntimeConfig", "2.0.0")
   type Platform = RuntimeConfig
-  @deprecated("use RuntimeConfig", "2.0.0")
-  val Platform = RuntimeConfig
+  // @deprecated("use RuntimeConfig", "2.0.0")
+  // val Platform = RuntimeConfig
 
   /**
    * Returns an effect that models success with the specified value.
